[REFACTOR] #419: 모델 관련 클래스 도메인 패키지로 묶기
