+++ backend/src/main/java/autoever2/cartag/models/ModelService.java
@@ -1,10 +1,11 @@
-package autoever2.cartag.service;
+package autoever2.cartag.models;
 
 import autoever2.cartag.domain.model.*;
 import autoever2.cartag.exception.EmptyDataException;
 import autoever2.cartag.exception.ErrorCode;
 import autoever2.cartag.cars.CarRepository;
-import autoever2.cartag.repository.ModelRepository;
+import autoever2.cartag.models.dto.*;
+import autoever2.cartag.models.ModelRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
