+++ backend/src/main/java/autoever2/cartag/models/dto/PowerTrainDataDto.java
@@ -1,4 +1,4 @@
-package autoever2.cartag.domain.model;
+package autoever2.cartag.models.dto;
 
 import lombok.Builder;
 import lombok.Getter;
