+++ backend/src/main/java/autoever2/cartag/service/QuoteService.java
@@ -6,7 +6,7 @@ import autoever2.cartag.cars.dto.TrimInfoDto;
 import autoever2.cartag.domain.quote.HistoryTotalModelPriceDto;
 import autoever2.cartag.domain.color.InnerColorDto;
 import autoever2.cartag.domain.color.OuterColorDto;
-import autoever2.cartag.domain.model.ModelDefaultDto;
+import autoever2.cartag.models.dto.ModelDefaultDto;
 import autoever2.cartag.domain.option.QuoteSubOptionDto;
 import autoever2.cartag.domain.option.SubOptionIdAndPriceDto;
 import autoever2.cartag.domain.quote.HistorySearchDto;
@@ -18,7 +18,7 @@ import autoever2.cartag.exception.ErrorCode;
 import autoever2.cartag.exception.InvalidDataException;
 import autoever2.cartag.recommend.RecommendConnector;
 import autoever2.cartag.repository.ColorRepository;
-import autoever2.cartag.repository.ModelRepository;
+import autoever2.cartag.models.ModelRepository;
 import autoever2.cartag.repository.OptionRepository;
 import autoever2.cartag.repository.QuoteRepository;
 import lombok.RequiredArgsConstructor;
