+++ backend/src/test/java/autoever2/cartag/repository/ModelRepositoryTest.java
@@ -1,7 +1,11 @@
 package autoever2.cartag.repository;
 
 import autoever2.cartag.domain.model.*;
-import autoever2.cartag.exception.EmptyDataException;
+import autoever2.cartag.models.ModelRepository;
+import autoever2.cartag.models.dto.ModelDefaultDto;
+import autoever2.cartag.models.dto.ModelDetailMappedDto;
+import autoever2.cartag.models.dto.ModelEfficiencyDataDto;
+import autoever2.cartag.models.dto.ModelShortMappedDto;
 import org.assertj.core.api.SoftAssertions;
 import org.assertj.core.api.junit.jupiter.InjectSoftAssertions;
 import org.assertj.core.api.junit.jupiter.SoftAssertionsExtension;
