+++ backend/src/test/java/autoever2/cartag/service/QuoteServiceTest.java
@@ -4,13 +4,12 @@ import autoever2.cartag.cars.CarRepository;
 import autoever2.cartag.cars.dto.TrimInfoDto;
 import autoever2.cartag.domain.color.InnerColorDto;
 import autoever2.cartag.domain.color.OuterColorDto;
-import autoever2.cartag.domain.model.ModelDefaultDto;
+import autoever2.cartag.models.dto.ModelDefaultDto;
 import autoever2.cartag.domain.option.QuoteSubOptionDto;
-import autoever2.cartag.domain.option.SubOptionIdAndPriceDto;
 import autoever2.cartag.domain.quote.*;
 import autoever2.cartag.recommend.RecommendConnector;
 import autoever2.cartag.repository.ColorRepository;
-import autoever2.cartag.repository.ModelRepository;
+import autoever2.cartag.models.ModelRepository;
 import autoever2.cartag.repository.OptionRepository;
 import autoever2.cartag.repository.QuoteRepository;
 import org.assertj.core.api.SoftAssertions;
